swarm: move and to its own module.
